# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sheep/{id}:
    get:
      operationId: getSheep
      tags: [sheep]
      summary: Get a single sheep by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sheep found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sheep"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Sheep:
      type: object
      required:
        [id, name, breed, ageMonths, weightKg, price, imageUrl, description, available, stockQuantity]
      properties:
        id:
          type: integer
        name:
          type: string
        breed:
          type: string
        ageMonths:
          type: integer
        weightKg:
          type: number
        price:
          type: number
        imageUrl:
          type: string
        description:
          type: string
        available:
          type: boolean
        stockQuantity:
          type: integer
...